+++ newrelic_api/tests/base_tests.py
@@ -1,5 +1,4 @@
 import os
-from datetime import time
 from unittest import TestCase
 
 from mock import patch, call, Mock
@@ -250,7 +249,7 @@ class ResourceTests(TestCase):
             0,
             [],
             {},
-            time(0)
+            None
         ]
 
         param_str = resource.build_param_string(test_params)