+++ newrelic_api/tests/mobile_applications_tests.py
@@ -3,67 +3,45 @@ from unittest import TestCase
 from mock import patch, Mock
 import requests
 
-from newrelic_api.applications import Applications
+from newrelic_api.mobile_applications import MobileApplications
 
 
-class NRApplicationsTests(TestCase):
+class NRMobileApplicationsTests(TestCase):
     def setUp(self):
-        super(NRApplicationsTests, self).setUp()
-        self.app = Applications(api_key='dummy_key')
+        super(NRMobileApplicationsTests, self).setUp()
+        self.app = MobileApplications(api_key='dummy_key')
 
         app_response = {
-            "id": 1234567,
-            "name": "demo_site",
-            "language": "python",
+            "id": 123456,
+            "name": "SampleApp",
             "health_status": "green",
             "reporting": True,
-            "last_reported_at": "2014-06-23T20:16:27+00:00",
-            "application_summary": {
-                "response_time": 170,
-                "throughput": 3,
-                "error_rate": 0,
-                "apdex_target": 0,
-                "apdex_score": 1
+            "mobile_summary": {
+                "active_users": 9792,
+                "launch_count": 9792,
+                "throughput": 71200,
+                "response_time": 0.428,
+                "calls_per_session": 7.27,
+                "interaction_time": 0.298,
+                "failed_call_rate": 1.2271918170353238,
+                "remote_error_rate": 0.10091960945281904
             },
-            "settings": {
-                "app_apdex_threshold": 0.5,
-                "end_user_apdex_threshold": 7,
-                "enable_real_user_monitoring": True,
-                "use_server_side_config": False
-            },
-            "links": {
-                "application_instances": [
-                    2345678,
-                    2345679
-                ],
-                "servers": [
-                    4567890,
-                    4567891
-                ],
-                "application_hosts": [
-                    5678901,
-                    5678902
-                ]
+            "crash_summary": {
+                "supports_crash_data": True,
+                "unresolved_crash_count": 0,
+                "crash_count": 0,
+                "crash_rate": 0
             }
         }
-        links = {
-            "application.servers": "/v2/servers?ids={server_ids}",
-            "application.server": "/v2/servers/{server_id}",
-            "application.application_hosts": "/v2/application/{application_id}/hosts?ids={host_ids}",
-            "application.application_host": "/v2/application/{application_id}/hosts/{host_id}",
-            "application.application_instances": "/v2/application/{application_id}/instances?ids={instance_ids}",
-            "application.application_instance": "/v2/application/{application_id}/instances/{instance_id}"
-        }
+
         self.list_success_response = {
             'applications': [
                 app_response,
-            ],
-            'links': links
+            ]
         }
 
         self.show_success_response = {
-            'application': app_response,
-            'links': links
+            'application': app_response
         }
 
         self.metric_name_response = {
@@ -123,7 +101,7 @@ class NRApplicationsTests(TestCase):
     @patch.object(requests, 'get')
     def test_list_success(self, mock_get):
         """
-        Test applications .list()
+        Test mobile_applications .list()
         """
         mock_response = Mock(name='response')
         mock_response.json.return_value = self.list_success_response
@@ -133,25 +111,6 @@ class NRApplicationsTests(TestCase):
         response = self.app.list()
 
         self.assertIsInstance(response, dict)
-
-    @patch.object(requests, 'get')
-    def test_list_success_with_filter_ids(self, mock_get):
-        """
-        Test applications .list()
-        """
-        mock_response = Mock(name='response')
-        mock_response.json.return_value = self.list_success_response
-        mock_get.return_value = mock_response
-
-        # Call the method
-        response = self.app.list(filter_ids=[1234567])
-
-        self.assertIsInstance(response, dict)
-        mock_get.assert_called_once_with(
-            url='https://api.newrelic.com/v2/applications.json',
-            headers=self.app.headers,
-            params='filter[ids]=1234567'
-        )
 
     @patch.object(requests, 'get')
     def test_list_failure(self, mock_get):
@@ -192,67 +151,6 @@ class NRApplicationsTests(TestCase):
         with self.assertRaises(ValueError):
             # Call the method
             self.app.show(id=1234567)
-
-    @patch.object(requests, 'get')
-    @patch.object(requests, 'put')
-    def test_update_success(self, mock_put, mock_get):
-        """
-        Test applications .update() success
-        """
-        mock_response = Mock(name='response')
-        mock_response.json.return_value = self.show_success_response
-        mock_get.return_value = mock_response
-        mock_put.return_value = mock_response
-
-        # Call the method
-        response = self.app.update(id=1234567, name='New Name')
-
-        self.assertIsInstance(response, dict)
-
-    @patch.object(requests, 'get')
-    @patch.object(requests, 'put')
-    def test_update_failure(self, mock_put, mock_get):
-        """
-        Test applications .update() failure
-        """
-        mock_response1 = Mock(name='response')
-        mock_response1.json.return_value = self.show_success_response
-        mock_get.return_value = mock_response1
-
-        mock_response = Mock(name='response')
-        mock_response.json.side_effect = ValueError('No JSON object could be decoded')
-        mock_put.return_value = mock_response
-
-        with self.assertRaises(ValueError):
-            # Call the method
-            self.app.update(id=1234567)
-
-    @patch.object(requests, 'delete')
-    def test_delete_success(self, mock_delete):
-        """
-        Test applications .delete() success
-        """
-        mock_response = Mock(name='response')
-        mock_response.json.return_value = self.show_success_response
-        mock_delete.return_value = mock_response
-
-        # Call the method
-        response = self.app.delete(id=1234567)
-
-        self.assertIsInstance(response, dict)
-
-    @patch.object(requests, 'delete')
-    def test_delete_failure(self, mock_delete):
-        """
-        Test applications .delete() failure
-        """
-        mock_response = Mock(name='response')
-        mock_response.json.side_effect = ValueError('No JSON object could be decoded')
-        mock_delete.return_value = mock_response
-
-        with self.assertRaises(ValueError):
-            # Call the method
-            self.app.delete(id=1234567)
 
     @patch.object(requests, 'get')
     def test_metric_names(self, mock_get):